fix: Linux did not compile, and now its decoder output is proven identical to Windows
The Linux build was broken, demonstrated rather than reasoned. In a clean
ubuntu:24.04 container with GCC 13:

  platform_linux.c:735:37: error: 'PTHREAD_MUTEX_INITIALIZER' undeclared here

plus implicit declarations of pthread_create, pthread_join, pthread_mutex_lock
and pthread_mutex_unlock. The file used all five with no <pthread.h>; glibc's
<sys/types.h> supplies the pthread TYPES, which is why this survived review,
but the macro has no other home. platform_macos.m includes it correctly. One
line fixes it, and nothing in the repository would ever have noticed, because
there is no CI and the tree was reorganised after the last Linux build.

CROSS-PLATFORM DECODE IS NOW A TEST, not a hope. tests/linux-build.sh builds in
a container and runs `dump` with no X server - it never calls graphicsInit and
touches no RNG - then compares every hash against the manifest generated on
Windows:

  windows: 2452 file(s)    linux: 2452 file(s)
  OK   every one of the 2452 decoded files is byte-identical to Windows

Two real defects had to be fixed to get there, and both were invisible before
anything compared the two platforms.

dump.c opened its output in TEXT mode, so the Windows CRT expanded every \n to
\r\n: 369 CR bytes in one 226 KB XPM, and all 2,452 files differing from Linux
for that reason alone. Now "wb". This changes Windows dump output, so the golden
manifest was regenerated deliberately.

The manifest itself had the same disease. Set-Content emits CRLF on Windows
PowerShell, so all 2,452 lines carried a trailing \r and every one differed when
read inside the container. That looked exactly like a catastrophic decoder
divergence and was purely line endings; the second time I chased it I checked
one file's hash on both sides first, found them identical, and knew the data was
fine before touching anything. Written with explicit LF now, and .gitattributes
pins eol=lf for the manifest and for *.sh, because core.autocrlf would otherwise
reintroduce both problems on a fresh checkout and additionally break the script
under bash.

Three more Linux defects fixed while in the file:

- snd_pcm_open succeeded, snd_pcm_hw_params failed, snd_pcm_close ran and the
  handle was left dangling. soundInit sets soundDisabled and carries on, so a
  later platformCloseAudio called snd_pcm_drain on freed memory by the ordinary
  path, not an exotic one.
- snd_pcm_writei's return was discarded, so the first XRUN - any scheduling
  hiccup - left the stream in an error state and every later write returned
  -EPIPE immediately: a 100% CPU spin with silent audio, nothing logged, for the
  life of the process. Now recovers via snd_pcm_recover and stops feeding if it
  cannot.
- malloc and pthread_create returns were unchecked, so a failure reported
  success to the caller with the device open and no thread feeding it.

ONE AUDIT FINDING REJECTED, recorded so it is not re-raised. The platform audit
reported that Linux and macOS call the audio callback without holding
audioMutex, making soundPlay's update unsynchronised. Checked: platformLockAudio
on both locks a real pthread mutex, and soundCallback takes it ITSELF
(sound.c:61), as does soundPlay. The critical section is inside the callback and
the state is properly synchronised. Worse, "fixing" it would deadlock on the
first sample: audioMutex is a default pthread mutex and is not recursive.
Windows only gets away with wrapping the callback because CRITICAL_SECTION is.

Two pre-existing -Wsign-compare warnings fixed (ttm.c, dump.c): the ternary had
an int and a size_t branch, so its common type was unsigned. Linux now builds
with no warnings at all under -Wall -Wpedantic -Wextra -Wshadow.

Windows gate still green: 12 smoke passed, 2,452 files byte-identical.

+++ src/engine/ttm.c
@@ -197,7 +197,9 @@ void ttmPlay(struct TTtmThread *ttmThread)     // TODO
             if ((i & 0x01) == 0x01)   // always read an even number of uint8s
                 strArg[i++] = (char)data[offset++];
 
-            strArg[i < (int)sizeof(strArg) ? i : sizeof(strArg) - 1] = '\0';
+            /* Both branches int: the second was size_t, so the ternary's common
+             * type was unsigned and GCC warned that `i` changed signedness. */
+            strArg[i < (int)sizeof(strArg) ? i : (int)sizeof(strArg) - 1] = '\0';
         }
         else {                        // args are numArgs words
             peekUint16Block(data, &offset, args, numArgs);
